Alloc when ExtractList
After reflect obtains ObjectList through List, it will call ExtractList
to get the Object list (internally, go-reflect is used to obtain the
pointer of the element in Items, because the pointer of this type is
the implementer of runtime.Object) and stored in Indexer. Suppose
there is only one Object in an ObjectList with a length of 1000 that
does not change, and other Objects are subsequently deleted, which
means that there is only one Object in the current Indexer, but because
the element address is directly taken here, the memory occupied by the
entire ObjectList is still Unable to release. Although most of objects
on this list is out of date.

We recreate the same type of Object through reflect, and set its fields
to disassociate the Object stored by the Indexer from the ObjectList.
This allows our program to quickly release the memory occupied by
obsolete Objects